Add some shared snippets
