+++ auth-next/index.js
@@ -70,3 +70,45 @@ function authStateListener() {
   });
   // [END auth_state_listener]
 }
+
+function setLanguageCode() {
+  // [START auth_set_language_code]
+  const { getAuth } = require("firebase/auth");
+
+  const auth = getAuth(firebaseApp);
+  auth.languageCode = 'it';
+  // To apply the default browser preference instead of explicitly setting it.
+  // firebase.auth().useDeviceLanguage();
+  // [END auth_set_language_code]
+}
+
+function authWithCredential(credential) {
+  // [START auth_signin_credential]
+  const { getAuth, signInWithCredential } = require("firebase/auth");
+
+  // Sign in with the credential from the user.
+  const auth = getAuth(firebaseApp);
+  signInWithCredential(auth, credential)
+    .then((result) => {
+      // Signed in 
+      // ...
+    })
+    .catch((error) => {
+      // Handle Errors here.
+      const errorCode = error.code;
+      const errorMessage = error.message;
+      // The email of the user's account used.
+      const email = error.email;
+      // ...
+    });
+  // [END auth_signin_credential]
+}
+
+function signInRedirect(provider) {
+  // [START auth_signin_redirect]
+  const { getAuth, signInWithRedirect } = require("firebase/auth");
+
+  const auth = getAuth(firebaseApp);
+  signInWithRedirect(auth, provider);
+  // [END auth_signin_redirect]
+}
