+++ auth/phone-auth.js
@@ -9,14 +9,6 @@ const window = {
   recaptchaVerifier: undefined
 };
 
-function setLanguageCode() {
-  // [START auth_set_language_code]
-  firebase.auth().languageCode = 'it';
-  // To apply the default browser preference instead of explicitly setting it.
-  // firebase.auth().useDeviceLanguage();
-  // [END auth_set_language_code]
-}
-
 function recaptchaVerifierInvisible() {
   function onSignInSubmit() {
     // TODO(you): Implement
