+++ snippets/auth-next/index/auth_set_language_code.js
@@ -1,5 +1,5 @@
 // This snippet file was generated by processing the source file:
-// ./auth-next/phone-auth.js
+// ./auth-next/index.js
 //
 // To make edits to the snippets in this file, please edit the source
