+++ snippets/auth-next/index/auth_signin_credential.js
@@ -1,17 +1,17 @@
 // This snippet file was generated by processing the source file:
-// ./auth-next/facebook.js
+// ./auth-next/index.js
 //
 // To make edits to the snippets in this file, please edit the source
 
-// [START auth_facebook_signin_credential_modular]
-import { getAuth, signInWithCredential, FacebookAuthProvider } from "firebase/auth";
+// [START auth_signin_credential_modular]
+import { getAuth, signInWithCredential } from "firebase/auth";
 
-// Sign in with the credential from the Facebook user.
+// Sign in with the credential from the user.
 const auth = getAuth(firebaseApp);
 signInWithCredential(auth, credential)
   .then((result) => {
     // Signed in 
-    const credential = FacebookAuthProvider.credentialFromResult(result);
+    // ...
   })
   .catch((error) => {
     // Handle Errors here.
@@ -19,8 +19,6 @@ signInWithCredential(auth, credential)
     const errorMessage = error.message;
     // The email of the user's account used.
     const email = error.email;
-    // The AuthCredential type that was used.
-    const credential = FacebookAuthProvider.credentialFromError(error);
     // ...
   });
-// [END auth_facebook_signin_credential_modular]
+// [END auth_signin_credential_modular]
