+++ snippets/auth-next/index/auth_signin_redirect.js
@@ -1,0 +1,11 @@
+// This snippet file was generated by processing the source file:
+// ./auth-next/index.js
+//
+// To make edits to the snippets in this file, please edit the source
+
+// [START auth_signin_redirect_modular]
+import { getAuth, signInWithRedirect } from "firebase/auth";
+
+const auth = getAuth(firebaseApp);
+signInWithRedirect(auth, provider);
+// [END auth_signin_redirect_modular]
